requirements.txt
numpy
pwntools
pycryptodome